ui update
- change title
- change icon
- change suspend button location
