+++ thai-online-debate/src/pages/edit-profile-page/EditProfile.jsx
@@ -1,12 +1,66 @@
 import './EditProfile.css'
 import UserNavBar from '../../components/Navbar/UserNavBar'
 import profileImg from '../../assets/profile.png'
-import React ,{ useContext ,useRef,useState}from 'react'
+import React, { useContext, useState, useRef, Suspense } from "react";
 import { AuthContext } from '../../context/authContext';
 import { makeRequest } from '../../axios';
 import { userPicPath } from '../../userPicPath';
+import { useNavigate } from 'react-router-dom';
+import Swal from 'sweetalert2';
+
 function EditProfile() {
     const { currentUser } = useContext(AuthContext);
+    const navigator = useNavigate();
+    const logout_db = async () => {
+        await makeRequest.post("/auth/logout", "");
+        localStorage.removeItem("user");
+        localStorage.removeItem('downloadList');
+        navigator("/signin");
+    }
+      const suspendedUser = async () => {
+          try {
+            const res = await makeRequest.put(`/auth/suspend`);
+            Swal.fire({
+              icon: "success",
+              title: "ระงับบัญชีสําเร็จ",
+            })
+            logout_db()
+          } catch (err) {
+            Swal.fire({
+              icon: "error",
+              title: err.response.data,
+            })
+          }
+          
+      }
+      const handleEditPasswordPopupClose = () => {
+        setShowEditPasswordPopup(false);
+      };
+      const suspendedClick = () => {
+        Swal.fire({
+          title: "คุณต้องการระงับบัญชีผู้ใช้ใช่หรือไม่",
+          text: 'กรุณายืนยันการระงับบัญชี: กรอกคำว่า "ระงับบัญชี" หากต้องการระงับบัญชีนี้',
+          icon: "warning",
+          iconColor: "red",
+          input: "text",
+          inputAttributes: {
+            autocapitalize: "off",
+          },
+          showCancelButton: true,
+          confirmButtonText: "ยืนยัน",
+          confirmButtonColor: "red",
+          cancelButtonText: "ยกเลิก",
+          showLoaderOnConfirm: true,
+          preConfirm: (inputValue) => {
+            if (inputValue !== "ระงับบัญชี") {
+              Swal.showValidationMessage(`ข้อความที่กรอกไม่ตรงกับ "ระงับบัญชี"`);
+            } else {
+              suspendedUser();
+            }
+          },
+          allowOutsideClick: () => !Swal.isLoading(),
+        });
+      };
   return (
     <>
     <UserNavBar/>
@@ -50,6 +104,15 @@ function EditProfile() {
                 <div className="edit-profile-profile-label-row">
                     {/* <p className='edit-profile-data-label'>รหัสผ่าน</p> */}
                     {/* <button className='edit-profile-edit-button'>เปลี่ยนรหัสผ่าน</button> */}
+                </div>
+                <div className='edit-profile-suspend-row'>
+                <button
+                type="button"
+                onClick={suspendedClick}
+                className="edit-profile-suspend-button"
+                >
+                ระงับบัญชีผู้ใช้งาน
+                </button>
                 </div>
 
             </div>
